Fix displayed number of nodes allocated and used in job view
The number of allocated nodes was infered from metrics associated
with the job. There are cases where a node's slurm-job-exporter will report
no usage for a job, even if a job was allocated. This typically
happens when no process for that job was launched for at least 60 seconds.

This PR makes the allocated number reflects the number of nodes allocated
by Slurm and the used number, the usage value infered from prometheus metrics.